libmultipath: move sysfs checks out of scsi_ioctl_pathinfo
There are multiple checks in scsi_ioctl_pathinfo() that only use sysfs.
pathinfo() is never called with DI_SERIAL and not DI_SYSFS, so it's safe
to move them to scsi_sysfs_pathinfo(). The HCTL checking code is already
there, so remove the duplicate, and split the sysfs serial checking from
the ioctl serial checking.

Signed-off-by: Benjamin Marzinski <bmarzins@redhat.com>
Reviewed-by: Martin Wilck <mwilck@suse.com>