libmultipath: retry DI_IOCTL functions if they were skipped
pathinfo() shouldn't call any of the DI_IOCTL functions if path_state is
not PATH_UP (It only actually checked this for scsi_ioctl_pathinfo).
But since pathinfo() is rarely called with the DI_IOCTL flag, this
information might never be gathered if a path is added, and pathinfo is
first called, while it's offline. Make paths remember if they skipped
the DI_IOCTL calls because the path was offline, and try again on future
pathinfo() calls that don't set DI_NOIO.

Also make dm_setgeometry check all paths, since it's possible that
set_geometry() was skipped (in the past, the ioctls would just fail for
the offline device, with the same result).

Signed-off-by: Benjamin Marzinski <bmarzins@redhat.com>
Reviewed-by: Martin Wilck <mwilck@suse.com>