libdmmp: update time stamp of man pages
The man pages should have the same timestamp as the last change
in libdmmp.h, which was 216de53 ("multipath-tools: use common convention
for "header file included" macros") in August 2024.

Together with the previous patch, this should ensure that the man pages
will get the correct time stamp in the future.

Signed-off-by: Martin Wilck <mwilck@suse.com>
Acked-by: Yoann Congal <yoann.congal@smile.fr>
Reviewed-by: Benjamin Marzinski <bmarzins@redhat.com>