multipathd: disable queueing on invalid multipath devices
If the multipath configuration has changed and a multipath device is no
longer valid after a reconfigure, disable queueing on it before
attempting to delete it. This makes it more likely to be deleted and if
the device isn't valid, then it there is no defined no_path_retry value
for it.

Fixes: https://github.com/opensvc/multipath-tools/issues/113
Signed-off-by: Benjamin Marzinski <bmarzins@redhat.com>
Reviewed-by: Martin Wilck <mwilck@suse.com>