multipathd: always reset checker interval for failed paths
The are multiple places in the code that can set pp->state to
PATH_DOWN without resetting pp->checkint, most obviously when
a path is failed by update_multipath(). When this happens the
path will continue to be checked at its previous interval,
instead of switching back to conf->checkint. Fix these cases
by always resetting the pp->checkint to conf->checkint when
updating the state of a failed path.

Signed-off-by: Benjamin Marzinski <bmarzins@redhat.com>
Reviewed-by: Martin Wilck <mwilck@suse.com>