libmultipath: remove free_paths argument from free_multipath()
free_multipath() is now always called with free_paths == KEEP_PATHS.
Remove the argument and the code for freeing paths.

We must make sure that if we free a struct mpp, no references remain to
it in any of the paths that are referenced by it. However, this means
that we must always free struct multipath before freeing any paths
that it links to.

orphan_paths() clears possible references in other paths, which
are not referenced from mpp->pg and mpp->paths, which we do mostly
for safety. But orphan_paths() also handles the add_when_online
logic, and must thus be called before free_multipath() in remove_map().

Suggested-by: Benjamin Marzinski <bmarzins@redhat.com>
Signed-off-by: Martin Wilck <mwilck@suse.com>
Reviewed-by: Benjamin Marzinski <bmarzins@redhat.com>