Makefile: fix CI build errors with cmocka 2.0
cmocka 2.0.0 introduces a couple of changes that require version-dependent
code to be used. Add macros to determine the cmocka version. If the version
can't be determined, assume 1.1.0.

Use a different approach here than in the stable branch, because the fix
there is rather large. Instead of replacing all deprecated macros, just
treat the respective warnings as non-fatal by adding the compiler flag
-Wno-error=deprecated-declarations for the CI code.

cmocka 2.0.1 added a macro to suppress these warnings altogether, but we
should also be able to build against 2.0.0.

Fixes: https://github.com/opensvc/multipath-tools/issues/129
Signed-off-by: Martin Wilck <mwilck@suse.com>